Refactor IdleBehavior namespace for better organization
- Moved `IdleBehavior` and related classes to `entity.behavior.idle` namespace.
- Updated all references to use the new namespace for consistency.
