+++ entity/behavior/idle/IdleBehavior.cs
@@ -1,5 +1,5 @@
 using Godot;
 
-namespace RougeLiteGame.entity.behavior;
+namespace RougeLiteGame.entity.behavior.idle;
 
 [GlobalClass] public abstract partial class IdleBehavior : Behavior;
