+++ entity/behavior/reaction/ReactionBehavior.cs
@@ -4,16 +4,16 @@ namespace RougeLiteGame.entity.behavior;
 
 [GlobalClass] public abstract partial class ReactionBehavior : Behavior
 {
-    protected IdleBehavior DefaultBehavior { get; private set; }
+    protected idle.IdleBehavior DefaultBehavior { get; private set; }
 
     protected Entity Target { get; private set; }
     
-    public void DefaultBehaviorOverride(IdleBehavior defaultBehavior)
+    public void DefaultBehaviorOverride(idle.IdleBehavior defaultBehavior)
     {
         DefaultBehavior = defaultBehavior;
     }
     
-    public void React(Entity target, IdleBehavior defaultBehavior)
+    public void React(Entity target, idle.IdleBehavior defaultBehavior)
     {
         DefaultBehavior = defaultBehavior;
         Target = target;
